+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `BatchSendingsEndpoint.list` endpoint
 - Add `AccountSecurity.visibleAutosignRoleOption` in to resource
 - Add `AccountSecurity.visibleNoneSignerChannelOption` in to resource
+- Add `BatchSendingsEndpoint.stats` endpoint
+- Extend `BatchSending` resource
 
 ## [2.5.0] - 2024-10-14
 - Add `AccountBilling.automaticTagsPlacement`